+++ src/Type/Definition/index.ts
@@ -5,6 +5,7 @@ export * from './EventDispatcher.js';
 export * from './EventIdentifier.js';
 export * from './EventListener.js';
 export * from './EventListenerIdentifier.js';
+export * from './Logger.js';
 export * from './OptionalPromise.js';
 export * from './PriorityRegistry.js';
 export * from './Registry.js';