+++ test/Unit/Type/Definition/EventDispatcher.test.ts
@@ -9,19 +9,7 @@ import {
   validateEventIdentifierFromString,
   validateEventListenerIdentifierFromString,
 } from '../../../../src/Type/Definition';
-
-test('event dispatcher can be created', () => {
-  const eventDispatcher = new EventDispatcher();
-  expect(eventDispatcher).toBeTruthy();
-  expect(eventDispatcher.hasListeners(validateEventListenerIdentifierFromString('some-identifier'))).toBeFalsy();
-});
-
-test('dispatching event without corresponding event listener works', () => {
-  const eventDispatcher = new EventDispatcher();
-  const event = new Event(validateEventIdentifierFromString('event'));
-  eventDispatcher.dispatchEvent(event);
-  expect(event.isPropagationStopped()).toBeFalsy();
-});
+import { testLogger } from '../../TestLogger';
 
 function createAnonymousEventListener(calledValue: string, stopsEvent: boolean = false): EventListener {
   return (event: EventInterface): OptionalPromise<void> => {
@@ -34,8 +22,33 @@ function createAnonymousEventListener(calledValue: string, stopsEvent: boolean =
   };
 }
 
+function getEventDispatcher(): EventDispatcher {
+  return new EventDispatcher(testLogger);
+}
+
+test('event dispatcher can be created', () => {
+  const eventDispatcher = getEventDispatcher();
+  expect(eventDispatcher).toBeTruthy();
+  expect(eventDispatcher.hasListeners(validateEventListenerIdentifierFromString('some-identifier'))).toBeFalsy();
+});
+
+test('dispatching event without corresponding event listener works', () => {
+  const eventDispatcher = getEventDispatcher();
+  const event = new Event(validateEventIdentifierFromString('event'));
+  eventDispatcher.dispatchEvent(event);
+  expect(event.isPropagationStopped()).toBeFalsy();
+});
+
+test('dispatching already stopped event leads to early return', () => {
+  const eventDispatcher = getEventDispatcher();
+  const event = new Event(validateEventIdentifierFromString('event'));
+  event.stopPropagation();
+  eventDispatcher.dispatchEvent(event);
+  expect(event.isPropagationStopped()).toBeTruthy();
+});
+
 test('dispatching event with single sync event listener works', async () => {
-  const eventDispatcher = new EventDispatcher();
+  const eventDispatcher = getEventDispatcher();
   const event = new Event(validateEventIdentifierFromString('event'), { called: [] });
   eventDispatcher.addListener(
     validateEventListenerIdentifierFromString('event'),
@@ -48,7 +61,7 @@ test('dispatching event with single sync event listener works', async () => {
 });
 
 test('dispatching event with single async event listener works', async () => {
-  const eventDispatcher = new EventDispatcher();
+  const eventDispatcher = getEventDispatcher();
   const event = new Event(validateEventIdentifierFromString('event'), { called: [] });
   eventDispatcher.addListener(
     validateEventListenerIdentifierFromString('event'),
@@ -69,7 +82,7 @@ test('dispatching event with single async event listener works', async () => {
 });
 
 test('priority of event listeners is respected', async () => {
-  const eventDispatcher = new EventDispatcher();
+  const eventDispatcher = getEventDispatcher();
   const event = new Event(validateEventIdentifierFromString('event'), { called: [] });
   eventDispatcher.addListener(
     validateEventListenerIdentifierFromString('event'),
@@ -93,7 +106,7 @@ test('priority of event listeners is respected', async () => {
 });
 
 test('indirect registration priority of event listeners is respected', async () => {
-  const eventDispatcher = new EventDispatcher();
+  const eventDispatcher = getEventDispatcher();
   const event = new Event(validateEventIdentifierFromString('event'), { called: [] });
   eventDispatcher.addListener(
     validateEventListenerIdentifierFromString('event'),
@@ -114,7 +127,7 @@ test('indirect registration priority of event listeners is respected', async () 
 });
 
 test('priority of event listener identifiers is respected', async () => {
-  const eventDispatcher = new EventDispatcher();
+  const eventDispatcher = getEventDispatcher();
   const event = new Event(validateEventIdentifierFromString('a.b.c'), { called: [] });
   eventDispatcher.addListener(
     validateEventListenerIdentifierFromString('a.b.c'),
@@ -133,7 +146,7 @@ test('priority of event listener identifiers is respected', async () => {
 });
 
 test('stoppable stops priority chain', async () => {
-  const eventDispatcher = new EventDispatcher();
+  const eventDispatcher = getEventDispatcher();
   const event = new Event(validateEventIdentifierFromString('a.b.c'), { called: [] });
   eventDispatcher.addListener(
     validateEventListenerIdentifierFromString('a.b.c'),
@@ -149,4 +162,134 @@ test('stoppable stops priority chain', async () => {
   await eventDispatcher.dispatchEvent(event);
   expect(event.isPropagationStopped()).toBeTruthy();
   expect(event.getContextValue('called')).toEqual(['first', 'second']);
+});
+
+test('exception in chain does not crash event dispatcher as a whole', async () => {
+  const eventDispatcher = getEventDispatcher();
+  const event = new Event(validateEventIdentifierFromString('test'), { called: [] });
+  eventDispatcher.addListener(
+    validateEventListenerIdentifierFromString('test'),
+    createAnonymousEventListener('first'),
+    250,
+  );
+  eventDispatcher.addListener(
+    validateEventListenerIdentifierFromString('test'),
+    (): OptionalPromise<void> => {
+      throw new Error('some error');
+    },
+    200,
+  );
+  eventDispatcher.addListener(
+    validateEventListenerIdentifierFromString('test'),
+    createAnonymousEventListener('third'),
+    150,
+  );
+
+  await eventDispatcher.dispatchEvent(event);
+  expect(event.isPropagationStopped()).toBeFalsy();
+  expect(event.getContextValue('called')).toEqual(['first', 'third']);
+});
+
+test('error in promise in chain does not crash event dispatcher as a whole', async () => {
+  const eventDispatcher = getEventDispatcher();
+  const event = new Event(validateEventIdentifierFromString('test'), { called: [] });
+  eventDispatcher.addListener(
+    validateEventListenerIdentifierFromString('test'),
+    createAnonymousEventListener('first'),
+    250,
+  );
+  eventDispatcher.addListener(
+    validateEventListenerIdentifierFromString('test'),
+    (): OptionalPromise<void> => {
+      return new Promise((_resolve, reject) => {
+        reject(new Error('Some error'));
+      });
+    },
+    200,
+  );
+  eventDispatcher.addListener(
+    validateEventListenerIdentifierFromString('test'),
+    createAnonymousEventListener('third'),
+    150,
+  );
+
+  await eventDispatcher.dispatchEvent(event);
+  expect(event.isPropagationStopped()).toBeFalsy();
+  expect(event.getContextValue('called')).toEqual(['first', 'third']);
+});
+
+test('get listeners returns empty array if no listeners are found', () => {
+  const eventDispatcher = getEventDispatcher();
+
+  const eventListeners = eventDispatcher.getListeners(validateEventListenerIdentifierFromString('i.do.not.exist'));
+  expect(eventListeners).toHaveLength(0);
+});
+
+test('get listeners returns registered event listeners', () => {
+  const eventListenerIdentifier = validateEventListenerIdentifierFromString('event');
+  const eventDispatcher = getEventDispatcher();
+  const firstEventListener = createAnonymousEventListener('first');
+  eventDispatcher.addListener(eventListenerIdentifier, firstEventListener);
+  const secondEventListener = createAnonymousEventListener('second');
+  eventDispatcher.addListener(eventListenerIdentifier, secondEventListener);
+
+  const eventListeners = eventDispatcher.getListeners(eventListenerIdentifier);
+  expect(eventListeners).toHaveLength(2);
+  expect(eventListeners).toContain(firstEventListener);
+  expect(eventListeners).toContain(secondEventListener);
+});
+
+test('get listeners does not return event listener with different identifier', () => {
+  const eventListenerIdentifier = validateEventListenerIdentifierFromString('event');
+  const otherEventListenerIdentifier = validateEventListenerIdentifierFromString('other');
+  const eventDispatcher = getEventDispatcher();
+  const firstEventListener = createAnonymousEventListener('first');
+  eventDispatcher.addListener(eventListenerIdentifier, firstEventListener);
+  const secondEventListener = createAnonymousEventListener('second');
+  eventDispatcher.addListener(otherEventListenerIdentifier, secondEventListener);
+
+  const eventListeners = eventDispatcher.getListeners(eventListenerIdentifier);
+  expect(eventListeners).toHaveLength(1);
+  expect(eventListeners).toContain(firstEventListener);
+});
+
+test('has listeners works', () => {
+  const eventDispatcher = getEventDispatcher();
+  const eventListenerIdentifier = validateEventListenerIdentifierFromString('event');
+  const eventListener = createAnonymousEventListener('some listener');
+
+  expect(eventDispatcher.hasListeners(eventListenerIdentifier)).toBeFalsy();
+  eventDispatcher.addListener(eventListenerIdentifier, eventListener);
+  expect(eventDispatcher.hasListeners(eventListenerIdentifier)).toBeTruthy();
+  eventDispatcher.removeListener(eventListenerIdentifier, eventListener);
+  expect(eventDispatcher.hasListeners(eventListenerIdentifier)).toBeFalsy();
+});
+
+test('remove listener does nothing if no event listeners were defined', () => {
+  const eventDispatcher = getEventDispatcher();
+  const eventListenerIdentifier = validateEventListenerIdentifierFromString('event');
+  const eventListener = createAnonymousEventListener('some listener');
+
+  expect(eventDispatcher.removeListener(eventListenerIdentifier, eventListener)).toBeTruthy();
+});
+
+test('remove listener does nothing if the deleted event listener is not found', () => {
+  const eventDispatcher = getEventDispatcher();
+  const eventListenerIdentifier = validateEventListenerIdentifierFromString('event');
+  const eventListener = createAnonymousEventListener('some listener');
+  const otherEventListener = createAnonymousEventListener('other listener');
+  eventDispatcher.addListener(eventListenerIdentifier, otherEventListener);
+
+  expect(eventDispatcher.removeListener(eventListenerIdentifier, eventListener)).toBeTruthy();
+});
+
+test('remove listener does remove registered event listener', () => {
+  const eventDispatcher = getEventDispatcher();
+  const eventListenerIdentifier = validateEventListenerIdentifierFromString('event');
+  const eventListener = createAnonymousEventListener('some listener');
+
+  eventDispatcher.addListener(eventListenerIdentifier, eventListener);
+  expect(eventDispatcher.hasListeners(eventListenerIdentifier)).toBeTruthy();
+  eventDispatcher.removeListener(eventListenerIdentifier, eventListener);
+  expect(eventDispatcher.hasListeners(eventListenerIdentifier)).toBeFalsy();
 });